move handle into database triggers to signal where the test cases are
