+++ packages/lix-sdk/src/own-entity-change-control/database-triggers.ts
@@ -1,8 +1,11 @@
 import type { SqliteDatabase } from "sqlite-wasm-kysely";
-import type { LixDatabaseSchema } from "../database/schema.js";
-import { handleLixOwnEntityChange } from "./handle-lix-own-entity-change.js";
+import type { LixDatabaseSchema, Snapshot } from "../database/schema.js";
 import type { Kysely } from "kysely";
-import { changeControlledTables } from "./change-controlled-tables.js";
+import {
+	changeControlledTables,
+	entityIdForRow,
+} from "./change-controlled-tables.js";
+import { createChange } from "../change/create-change.js";
 
 export function applyOwnEntityChangeControlTriggers(
 	sqlite: SqliteDatabase,
@@ -50,5 +53,72 @@ export function applyOwnEntityChangeControlTriggers(
       `;
 
 		sqlite.exec(sql);
+	}
+}
+
+async function handleLixOwnEntityChange(
+	db: Kysely<LixDatabaseSchema>,
+	tableName: keyof LixDatabaseSchema,
+	operation: "insert" | "update" | "delete",
+	...values: any[]
+): Promise<void> {
+	const executeInTransaction = async (trx: Kysely<LixDatabaseSchema>) => {
+		// need to break the loop if own changes are detected
+		const change = await trx
+			.selectFrom("change")
+			.where("id", "=", values[0])
+			.select("plugin_key")
+			.executeTakeFirst();
+
+		if (change?.plugin_key === "lix_own_entity") {
+			return;
+		}
+		const currentVersion = await trx
+			.selectFrom("current_version")
+			.innerJoin("version", "current_version.id", "version.id")
+			.selectAll("version")
+			.executeTakeFirstOrThrow();
+
+		const authors = await trx
+			.selectFrom("active_account")
+			.selectAll()
+			.execute();
+
+		if (authors.length === 0) {
+			console.error(tableName, change);
+			throw new Error("At least one author is required");
+		}
+
+		let snapshotContent: Snapshot["content"] | null;
+
+		if (operation === "delete") {
+			snapshotContent = null;
+		} else {
+			snapshotContent = {};
+			// construct the values as json for the snapshot
+			for (const [index, column] of changeControlledTables[
+				tableName
+			]!.entries()) {
+				snapshotContent[column] = values[index];
+			}
+		}
+
+		const entityId = entityIdForRow(tableName, values);
+
+		await createChange({
+			lix: { db: trx },
+			authors: authors,
+			version: currentVersion,
+			entityId,
+			fileId: "null",
+			pluginKey: "lix_own_entity",
+			schemaKey: `lix_${tableName}`,
+			snapshotContent,
+		});
+	};
+	if (db.isTransaction) {
+		await executeInTransaction(db);
+	} else {
+		await db.transaction().execute(executeInTransaction);
 	}
 }
